threads use animations instead of commands

+++ lib/Animation.hpp
@@ -17,7 +17,7 @@ protected:
 
     LocalStack& stack;
 
-    int updateRate;
+    unsigned long int updateRate;
     int max_exec;
     int current_exec;
